fix(dashboard): address PR review on leaderboard
Logic fixes:
- rangeLabel: restore "24h" / "7d" branches (cursor + codex). Headings
  used to show "Top traders (all-time)" for short windows even though
  the underlying queries were filtered to the selected range.
- rangeSubtitle: restore "Today (UTC)" for "24h" — was rendering
  "Last 1 days" (cursor).
- Cap banner: include aggregator-only branch — previously fell through
  to "pool-day" when only `isV2AggregatorCapHit` was set (cursor).
- Fallback / single-line chart pill set: pass new
  `LEADERBOARD_FALLBACK_CHART_RANGES` (1W/1M/3M/All) so v2+90d (or any
  fallback case) has an active pill — defaulted to RANGES (1W/1M/All)
  before, which doesn't include 3M (cursor + codex).

File-size splits (codex P2 — both files were over the AGENTS.md 600-line
soft cap):
- `lib/leaderboard.ts` (692 → 485): per-pool aggregation +
  PoolDailyVolumeRow / PoolBreakdown moved to `lib/leaderboard-pool.ts`
  (216 lines). Re-exported for backward compat.
- `app/leaderboard/page-client.tsx` (829 → 576): URL-state machinery
  (range / system toggle / venue + UTC-day rollover ticker) moved to
  `_lib/url-state.ts` (175). Pool-chart view-model (allowlist, window
  range, breakdown decoration, Top Pools list rows) moved to
  `_lib/pool-chart-vm.tsx` (157).
